Made proposal description markdown scroll.
