+++ packages/stateless/components/index.tsx
@@ -28,7 +28,7 @@ export * from './GridCardContainer'
 export * from './HorizontalNftCard'
 export * from './LineGraph'
 export * from './LinkWrapper'
-export * from './MarkdownPreview'
+export * from './MarkdownRenderer'
 export * from './NftCard'
 export * from './NoContent'
 export * from './StatusDisplay'
